普通码应该是用不了，即使抬离纸面也会触发OIDRECEIVE

+++ src/Global_vars.h
@@ -13,5 +13,5 @@ extern int OID_Y;
 extern int OID_Angle;
 extern boolean OID_available;
 extern int OID_Index;
-extern int OID_code_type;
+extern uint8_t OID_code_type;
